Extract AdditionalSpecifications
