+++ default/python/AI/ship_design/__init__.py
@@ -1,0 +1,1 @@
+from ship_design._additional_specification import AdditionalSpecifications
